xbd_pixel_baseline_prior/requirements.txt
torch
torchvision
Pillow
numpy
PyYAML
tqdm
matplotlib
